requirements.txt
# LangChain e LLMs
langchain==0.1.17
langchain-core==0.1.34
langchain-community==0.0.21
langchain-openai==0.1.7
langchain-huggingface==0.0.1
openai==1.16.1

# Transformers e HuggingFace
transformers==4.39.3
accelerate==0.29.3
sentence-transformers==2.6.1  # necessário para embeddings

# Streamlit e mapas
streamlit==1.32.2
streamlit-folium==0.19.0

# Utilitários
httpx==0.27.2
jupyter
pandas
python-dotenv

# Compatibilidade
protobuf<5
packaging<24
numpy<2.0
